create PollForm and adapt Create and EditPollModal
