+++ js/src/common/index.js
@@ -1,0 +1,1 @@
+export const slug = 'fof-polls';
