+++ js/src/forum/components/form/FormError.js
@@ -1,0 +1,3 @@
+export default class FormError extends Error {
+
+}
