+++ js/src/forum/components/index.ts
@@ -2,10 +2,12 @@ import CreatePollModal from './CreatePollModal';
 import PostPoll from './PostPoll';
 import EditPollModal from './EditPollModal';
 import ListVotersModal from './ListVotersModal';
+import PollForm from './PollForm';
 
 export const components = {
   CreatePollModal,
   PostPoll,
   EditPollModal,
   ListVotersModal,
+  PollForm,
 };
