+++ js/src/forum/states/PollFormState.js
@@ -1,0 +1,58 @@
+import app from 'flarum/forum/app';
+
+export default class PollFormState {
+    constructor(poll) {
+        if (!poll) {
+            poll = app.store.createRecord('polls');
+            poll.pushAttributes( {
+                question: '',
+                endDate: '',
+                publicPoll: false,
+                allowMultipleVotes: false,
+                hideVotes: false,
+                allowChangeVote: false,
+                maxVotes: 0,
+            });
+            poll.pushData({relationships: {options: []}});
+        }
+
+        this.loading = false;
+        this.deleting = false;
+        this.poll = poll;
+        this.expandedGroup = 'setup';
+    }
+
+    isExpanded(groupKey) {
+        return this.expandedGroup === groupKey;
+    }
+
+    expand(groupKey) {
+        this.expandedGroup = groupKey;
+        m.redraw();
+    }
+
+    async save(data) {
+        this.loading = true;
+        m.redraw();
+
+        try {
+            this.poll = await this.poll.save(data);
+        } finally {
+            this.loading = false;
+            m.redraw();
+        }
+    }
+
+    async delete() {
+        this.loading = true;
+        m.redraw();
+
+        try {
+            await this.poll.delete();
+            this.deleting = true;
+        } finally {
+            this.loading = false;
+            m.redraw();
+        }
+    }
+}
